pyproject.toml
[project]
name = "snip-mcp"
version = "0.1.0"
description = "Screen snipping MCP server - capture screen regions as images in conversation context"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
snip-mcp = "snip_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/snip_mcp"]
